personal_copilit_openai_SDK/requirements.txt
python-dotenv
openai
openai-agents
gradio